book Cricket game
The Game of Book Cricket is played by 2 players.
Each player randomly opens the book and as you open the pages you score runs. 
 
Constraints :
Rules for the Game:

Assume a book of 300 pages
Player 1 opens the book to read a page number (Randomly generated)
Total score of a player = page number % 7 (number % 7 will always be between 0 - 6)
Player 1 continues and keeps adding to the score untill page number results in a 0 score
Player 2 starts the game and tries to beat the score of player 1
In case of a tie, check the number of turns(player with lesser turns is winner)
Show the winner of the game.